+++ modules/S14_weak_func_radare_check.sh
@@ -419,7 +419,9 @@ radare_output_function_details()
   LOG_FILE_LOC_OLD="$LOG_FILE_LOC"
   LOG_FILE_LOC="$LOG_PATH_MODULE"/vul_func_"$COUNT_FUNC"_"$FUNCTION"-"$NAME".txt
 
-  mv "$LOG_FILE_LOC_OLD" "$LOG_FILE_LOC" 2> /dev/null
+  if [[ -f "$LOG_FILE_LOC_OLD" ]]
+    mv "$LOG_FILE_LOC_OLD" "$LOG_FILE_LOC" 2> /dev/null
+  fi
   
   if [[ "$NETWORKING" -gt 1 ]]; then
     NETWORKING_="${ORANGE}networking: yes${NC}"
